\name{CreateInfo}
\alias{CreateInfo}
\title{Create info list function for popPK}
\description{Creates info list for popPK data.}
\usage{
	 CreateInfo(info)
}
\arguments{\item{info}{See \link{popPK}}}

\value{
  Creates info list with all required arguments.
}
\note{
  For examples see \link{popPK}.
  }

\author{Christoffer W. Tornoe & Fang Li \email{pharmacometrics@fda.hhs.gov}}

\keyword{models}
