\name{Demographics}
\alias{Demographics}
\title{Demographics function for popPK}
\description{Creates summary demographic tables and graphs.}
\usage{
	 Demographics(info)
}
\arguments{\item{info}{See \link{popPK}}}

\value{
  Creates tables (CatCovariates and ConCovariates), histograms (CovHist, CatHist),
  and scatter plot matrices of continuous covariates (CovSplom).
}
\note{
  For examples see \link{popPK}.
  }

\author{Christoffer W. Tornoe & Fang Li \email{pharmacometrics@fda.hhs.gov}}

\keyword{models}
