\name{printGraph}
\alias{printGraph}
\title{Function to print graphs using different devices}
\description{Creates info list for popPK data.}
\usage{
	 printGraph(name,device,height,width)
}
\arguments{
	\item{name}{File name of graph.}
	\item{device}{Graphical device to print to (options: "wmf", "bmp", "jpg", "pdf", "png", or "tiff") (default is "pdf").}
	\item{height}{Height of graph.}
	\item{width}{Width of graph.}
}

\note{
  For examples see \link{popPK}.
}

\author{Christoffer W. Tornoe & Fang Li \email{pharmacometrics@fda.hhs.gov}}

\keyword{models}
