\name{xposeData}
\alias{xposeData}
\title{Import data for popPK}
\description{Import NONMEM run using xpose.data function in xpose4 for popPK.}
\usage{
	 xposeData(info)
}
\arguments{\item{info}{See \link{popPK}}}

\value{
  \item{xpdb}{xposeData returns the xpose.data object xpdb.}
}
\note{
  For examples see \link{popPK}.
}

\author{Christoffer W. Tornoe & Fang Li \email{pharmacometrics@fda.hhs.gov}}

\keyword{models}
